requirements.txt
flask
python-dotenv
requests
mangum # Para hacer el despliegue en lambdadeactivate